pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git-alias2"
version = "2.0.0"
description = "Git aliases for multi-provider repository management"
requires-python = ">=3.10"
dependencies = [
    "GitPython>=3.1.43",
    "raccoontools>=1.1.1",
]

[project.scripts]
g = "g:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools]
py-modules = ["g"]

[tool.setuptools.packages.find]
include = ["core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
